Modify functions to create and unpickle subsets of the dataset

+++ src/classifier/convert_data.py
@@ -5,16 +5,13 @@ We use the NIST SD19 by_class dataset
 import os
 import time
 import pickle
-import collections
 import argparse
 from PIL import Image
 import numpy as np
+import nist_by_class
 
 HEIGHT = 32
 WIDTH = 32
-LABELFILES = collections.namedtuple('Label_Files', ['Directory',
-                                                    'Training',
-                                                    'Testing'])
 
 
 def resize_images(directory):
@@ -25,7 +22,6 @@ def resize_images(directory):
     """
     start = time.time()
     dirs = os.listdir(directory)
-
     # Go through each label folder
     for label in dirs:
         label_path = os.path.join(directory, label)
@@ -42,22 +38,28 @@ def resize_images(directory):
                 image_path = os.path.join(data_path, item)
 
                 # Open the image and convert to grayscale
-                im = Image.open(image_path).convert('L')
+                image = Image.open(image_path).convert('L')
+                width, height = image.size
 
                 # Grab the filename
                 fname = os.path.splitext(image_path)[0] + '_' + label + '.png'
 
-                # Resize the image from 128 x 128 to 32 x 32
-                imResize = im.resize((WIDTH, HEIGHT), Image.LANCZOS)
+                if width != WIDTH | height != HEIGHT:
+                    # Resize the image from 128 x 128 to 32 x 32
+                    resized_image = image.resize((WIDTH, HEIGHT),
+                                                 Image.LANCZOS)
 
-                # Save the new image and replace the old image
-                imResize.save(fname, 'PNG')
+                    # Save the new image and replace the old image
+                    resized_image.save(fname, 'PNG')
 
-                # Print the progress
-                print('Resized ' + item)
+                    # Print the progress
+                    print('Resized ' + item)
 
-                # Remove the old image
-                os.remove(image_path)
+                    # Remove the old image
+                    os.remove(image_path)
+
+                else:
+                    print(item + ' is already the correct size')
 
     # Time
     print('> Completion Time: {}'.format(time.time() - start))
@@ -74,8 +76,8 @@ def convert_to_pixel_array(image_path):
     """
     pixels = []
 
-    im = Image.open(image_path)
-    pixels = list(im.getdata())
+    iamge = Image.open(image_path)
+    pixels = list(iamge.getdata())
 
     # Normalize and zero center pixel data
     std_dev = np.std(pixels)
@@ -96,145 +98,132 @@ def convert_to_pixel_array(image_path):
 def generate_label_list(directory):
     """
     Generate a text file that lists all of the image paths with their label
-    Input: by_class directory (after running nist_by_class script)
+    Input: NIST_split directory (after running nist_by_class script)
     Output: text file of image paths with their label
     """
-    os.chdir(directory)
-    root_dirc = os.listdir()
-
-    # Create the text filename
-    training_file_name = "char_training_label.txt"
-    test_file_name = "char_testing_label.txt"
-
-    # Create the label files
     start = time.time()
 
-    # Open the file, create if they do not exist
-    train_file = open(os.path.join(os.pardir, training_file_name), "w+")
-    test_file = open(os.path.join(os.pardir, test_file_name), "w+")
+    # Change to the directory
+    os.chdir(directory)
 
-    # Go through all of the label folders
-    for label in root_dirc:
-        label_path = os.path.join(directory, label)
-        os.chdir(label_path)
-        label_dirc = os.listdir()
+    # List out subset folders
+    root_dirc = os.listdir()
 
-        # Each label folder has two folders, training and testing
-        for folder in label_dirc:
-            # Write to the training file
-            if folder == 'training':
-                folder_path = os.path.join(label_path, folder)
-                os.chdir(folder_path)
-                folder_dirc = os.listdir()
-                for image in folder_dirc:
-                    image_path = os.path.join(folder_path, image)
-                    train_file.write(str(image) +
-                                     "," +
-                                     str(image_path) +
-                                     ","
-                                     + str(label)
-                                     + "\r\n")
+    for subset in root_dirc:
+        if os.path.isdir(subset):
+            label_path = os.path.join(directory, subset)
+            labels = os.listdir(label_path)
 
-            # Write to the testing file
-            if folder == 'testing':
-                folder_path = os.path.join(label_path, folder)
-                os.chdir(folder_path)
-                folder_dirc = os.listdir()
-                for image in folder_dirc:
-                    image_path = os.path.join(folder_path, image)
-                    test_file.write(str(image) +
-                                    "," +
-                                    str(image_path) +
-                                    ","
-                                    + str(label)
-                                    + "\r\n")
-    # Close the files
+            train_file = "char_train_label_" + str(subset) + ".txt"
+            test_file = "char_test_label_" + str(subset) + ".txt"
+
+            train_file = open(os.path.join(directory, train_file), "w+")
+            test_file = open(os.path.join(directory, test_file), "w+")
+
+            for label in labels:
+                folder_path = os.path.join(label_path, label)
+                folders = os.listdir(folder_path)
+
+                for folder in folders:
+                    image_path = os.path.join(folder_path, folder)
+                    images = os.listdir(image_path)
+
+                    for image in images:
+                        img_path = os.path.join(image_path, image)
+                        if folder == 'training':
+                            train_file.write(str(image) +
+                                             "," +
+                                             str(img_path) +
+                                             "," +
+                                             str(label) +
+                                             "\r\n")
+                        if folder == 'testing':
+                            test_file.write(str(image) +
+                                            "," +
+                                            str(img_path) +
+                                            "," +
+                                            str(label) +
+                                            "\r\n")
+
     train_file.close()
     test_file.close()
 
     # Time
     print('> Label files complete: {}'.format(time.time() - start))
-
-    # Return the root dataset directory, training file name, and
-    # testing file name
-    result = LABELFILES(directory, training_file_name, test_file_name)
-    return result
+    return directory
 
 
-def pickle_data(directory, text_file, datafile):
+def pickle_data(directory, text_file_no, datafile, setting):
     """
     Read a text file and pickle the contents to prevent work repeat
     Input: Text file containing the filename, absolute image path, label
     Output: Pickled file of the character data
     """
-    # Point is a set that contains the image file name, its path,
-    # the label, and the actual image pixel array.
-    point = {}
-
-    # Char_data is the list of all points from the label text file
-    char_data = []
-
     start = time.time()
+    # Point is a dictionary that contains the image file name, its path,
+    # the label, and the actual image pixel array.
+
+    # Char_data is the dictionary of all points from the label text file
+    char_data = []
 
     # First, change to the dataset root directory
     os.chdir(directory)
 
-    try:
-        # Open the text file
-        with open(text_file) as train:
-            for line in train:
-                # Remove the new line character from each line
-                v = line.strip().split(',')
+    # Find the text file
+    if setting == "training":
+        text_file = "char_train_label_subset_" + str(text_file_no) + ".txt"
+    elif setting == "testing":
+        text_file = "char_test_label_subset_" + str(text_file_no) + ".txt"
+    else:
+        print("We need the correct setting")
 
-                # Filename is the first element of the set
-                point['filename'] = v[0]
+    # Open the text file
+    if text_file:
+        try:
+            with open(text_file) as text:
+                for line in text:
+                    # Remove the new line character from each line
+                    text_line = line.strip().split(',')
+                    image = convert_to_pixel_array(text_line[1])
+                    text_line.append(image)
+                    char_data.append(text_line)
+            pickle.dump(char_data, open(datafile, "wb"),
+                        protocol=pickle.HIGHEST_PROTOCOL)
+        except FileNotFoundError:
+            print('Text file {} not foud'.format(text_file))
 
-                # Image path is the second element
-                point['image_path'] = v[1]
-
-                # Image label is the thrid element
-                point['label'] = v[2]
-
-                # Converted image pixel array is the last element of the set
-                point['pixel_array'] = convert_to_pixel_array(v[1])
-
-                # Append the set to the list
-                char_data.append(point)
-
-        # The complete list is then pickled into the specified filename
-        pickle.dump(char_data, open(datafile, "wb"),
-                    protocol=pickle.HIGHEST_PROTOCOL)
-        print("> Pickle time: {}".format(time.time() - start))
-
-        # Return the pickle file name
-        return datafile
-
-    # Throw an error if we can't find the text file
-    except FileNotFoundError:
-        print('Label file not found at path {}'.format(directory))
+    print("> Pickle time: {}".format(time.time() - start))
 
 
-def load_data(directory, datafile):
+def load_data(directory, datafile_no, setting):
     """
     Function to load a pickled datafile and return its contents
-    Input: Directory to the pickle file, Pickle filename
+    Input: Directory to the pickle file, subset no, setting(testing/training)
     Output: Contents of specified pickle file
     """
-    # Change to the specified directory
     os.chdir(directory)
 
-    try:
-        # Load the pickled file
-        data = pickle.load(open(datafile, "rb"))
+    # Based on setting, load the specified pkl file
+    if setting == "testing":
+        fname = "test_pickle_" + str(datafile_no) + ".pkl"
+    elif setting == "training":
+        fname = "train_pickle_" + str(datafile_no) + ".pkl"
+    else:
+        print("The setting must be specified as testing or training")
 
-        # Return the contents
-        print('Data loaded from {}'.format(datafile))
-        return data
+    # Unpickle the pickle
+    if fname:
+        try:
+            data = pickle.load(open(fname, "rb"))
+            print('Data loaded from {}'.format(fname))
+        except FileNotFoundError:
+            print('{} not found'.format(fname))
+            print('Creating {}'.format(fname))
+            pickle_data(directory, datafile_no, fname, setting)
+            data = pickle.load(open(fname, "rb"))
+            print('Data loaded from {}'.format(fname))
 
-    # Throw an error if we can't find the pickle file
-    except FileNotFoundError:
-        print('Pickle file {} not found at path {}'
-              .format(datafile, directory))
+    return data
 
 
 if __name__ == "__main__":
@@ -249,26 +238,23 @@ if __name__ == "__main__":
     #############################
 
     # First, resize the images
-    resize_path = resize_images(ARGS['directory'])
+    RESIZED = resize_images(ARGS['directory'])
 
-    # # Second, generate the label text files
-    labels_path = generate_label_list(resize_path)
+    # Second, spilt the dataset into multiple subsets
+    SPLIT, DIVISION = nist_by_class.split_NIST(RESIZED, 10)
 
-    # # Pickle the data
-    train_pickle = pickle_data(labels_path.Directory,
-                               labels_path.Training,
-                               "trainingg_set.p")
+    # Third, generate the label text files
+    DATA = generate_label_list(SPLIT)
 
-    test_pickle = pickle_data(labels_path.Directory,
-                              labels_path.Testing,
-                              "testing_set.p")
+    # Pickle the ten subsets
+    COUNT = 0
+    while COUNT < DIVISION:
+        TRAINNAME = "train_pickle_" + str(COUNT) + ".pkl"
+        TESTNAME = "test_pickle_" + str(COUNT) + ".pkl"
+        pickle_data(DATA, COUNT, TRAINNAME, "training")
+        pickle_data(DATA, COUNT, TESTNAME, "testing")
+        COUNT += 1
 
-    # Load the pickle data
-    train = load_data(labels_path.Directory, train_pickle)
-    test = load_data(labels_path.Directory, test_pickle)
-
-    for entry in train:
-        print(entry)
-
-    for entry in test:
-        print(entry)
+    # # Load the pickle data
+    TRAIN = load_data(DATA, 9, "training")
+    TEST = load_data(DATA, 9, "testing")
